requirements.txt
selenium
plyer
python-dotenv
requests
